+++ tests/integration/cases/int_list_single/Ada_combined@ada_2022.adb
@@ -1,0 +1,10 @@
+with A_Stub; use A_Stub;
+procedure Main is
+    my_data : A_Val := AList'[
+        AInt (1)
+    ];
+begin
+    my_data := AList'[
+        AInt (1)
+    ];
+end Main;