+++ tests/integration/cases/int_list_single/CSharp@v10.cs
@@ -1,0 +1,4 @@
+using System;
+var my_data = (
+    1
+);