+++ tests/integration/cases/int_list_single/Dart@v3.dart
@@ -1,0 +1,3 @@
+final my_data = <int>[
+    1,
+];