+++ tests/integration/cases/int_list_single/FSharp_combined@v8.fs
@@ -1,0 +1,16 @@
+module Main
+
+type Val =
+    | FInt of int64
+    | FList of Val list
+let private _mainDeclaration () =
+    let mutable my_data: Val = FList [
+        FInt 1L
+    ]
+    ignore my_data
+
+let private _mainAssignment () =
+    let my_data: Val = FList [
+        FInt 1L
+    ]
+    ignore my_data