+++ tests/integration/cases/int_list_single/Go@v1_18.go
@@ -1,0 +1,8 @@
+package main
+
+func main() {
+my_data := []int{
+	1,
+}
+_ = my_data
+}