+++ tests/integration/cases/int_list_single/JavaScript_combined@es2015.js
@@ -1,0 +1,6 @@
+let my_data = [
+  1,
+];
+my_data = [
+  1,
+];