+++ tests/integration/cases/int_list_single/Kotlin_sequence_tuple_decl_var@v1_9.kts
@@ -1,0 +1,3 @@
+var my_data = listOf<Any?>(
+    1,
+)