+++ tests/integration/cases/int_list_single/Python_combined@py39.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+my_data = (
+    1,
+)
+my_data = (
+    1,
+)