+++ tests/integration/cases/int_list_single/Rust_combined@edition_2021.rs
@@ -1,0 +1,9 @@
+fn main() {
+    let mut my_data = vec![
+        1,
+    ];
+    my_data = vec![
+        1,
+    ];
+    let _ = my_data;
+}