+++ tests/integration/cases/int_list_single/Swift_combined@v5_9.swift
@@ -1,0 +1,6 @@
+var my_data: Any = [
+    1,
+]
+my_data = [
+    1,
+]